+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta5] - In progress
+
+### Added
+
+- New method `PayloadInterface::getOrFail(string $path): mixed`
+
 ## [1.0.0-beta4] - 2025-01-20
 
 ### Added